Refactored useMockResponse code into mock.tsx
