+++ packages/framework/src/evaluate/binding.ts
@@ -88,7 +88,6 @@ export const createBindingAtom = <T = unknown>(
         },
         env: get(envAtom),
         secrets: get(secretAtom),
-        useMockResponse: useMockResponse(appData.application?.id)
       },
       screenContext: {
         inputs: get(screenInputAtom),
