+++ packages/runtime/src/runtime/invokeApi.tsx
@@ -6,6 +6,7 @@ import {
   type Response,
   type useScreenData,
 } from "@ensembleui/react-framework";
+import { mock } from "./mock";
 
 export const invokeAPI = async (
   screenData: ReturnType<typeof useScreenData>,
@@ -32,33 +33,8 @@ export const invokeAPI = async (
   });
 
   // Check to see if mockResponse is enabled and if a mockResponse exists in the API context
-  if (api.mockResponse && useMockResponse(appData?.application?.id)) {
-    // Ensure that the mock response contains a correctly formatted
-    if (typeof evaluatedMockResponse.response !== "object")
-      throw new Error(
-        "Improperly formatted mock response: Malformed mockResponse object",
-      );
-
-    if (
-      !evaluatedMockResponse.response.statusCode ||
-      typeof evaluatedMockResponse.response.statusCode !== "number"
-    )
-      throw new Error(
-        "Improperly formatted mock response: Incorrect Status Code. Please check that you have included a status code and that it is a number",
-      );
-
-    const isSuccess =
-      evaluatedMockResponse.response.statusCode >= 200 &&
-      evaluatedMockResponse.response.statusCode <= 299;
-    const mockRes = {
-      ...evaluatedMockResponse.response,
-      isLoading: false,
-      isSuccess,
-      isError: !isSuccess,
-    };
-    screenData.setData(api.name, mockRes);
-    return mockRes;
-  }
+  if (api.mockResponse && useMockResponse(appData?.application?.id))
+    return mock(evaluatedMockResponse);
 
   // If mock resposne does not exist, fetch the data directly from the API
   const res = await DataFetcher.fetch(api, { ...apiInputs, ...context });
